fix: allow empty selector
While unconventional, there may be no selector at all. In that case,
the constraints will always be on (and can't be turned off).
See the added test `no_selector` for an example.
However, halo2-analyzer instead skips the constraints completely,
causing a failure. This commit fixes the issue.

Note that this commit only includes the fix for the under-constrained
circuit analysis. Other types of analysis may still be incorrect,
and require further fixes.

The bug is originally discovered by @shankarapailoor.